started ticket forms

+++ src/main/java/com/web/TicketController.java
@@ -8,9 +8,14 @@ import com.service.SeatService;
 import com.service.TicketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.view.RedirectView;
 
+import javax.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -26,12 +31,15 @@ public class TicketController
     @Autowired
     private SeatService seatService;
 
-    @RequestMapping(value = "/admin/addTicket", method = RequestMethod.POST)
-    public @ResponseBody ModelAndView addTicket(@RequestBody Ticket ticket) throws Exception
+    @Autowired
+    private TicketEditor ticketEditor;
+
+    @Autowired
+    private FilmshowEditor filmshowEditor;
+
+    @RequestMapping("/admin/addTicket")
+    public ModelAndView addTicket(@ModelAttribute Ticket ticket) throws Exception
     {
-        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
-        ModelAndView mav = new ModelAndView("addTicket");
-        mav.addObject("filmshowListJson", filmshowList);
         List<Ticket> ticketList = ticketService.getTicketAll();
         boolean ticketValid = true;
         if(ticket != null && ticket.getFilmshow() != null
@@ -48,18 +56,18 @@ public class TicketController
             if(ticketValid)
             {
                 ticketService.create(ticket);
-                mav.addObject("ticketJson", ticket);
+                return new ModelAndView(new RedirectView("/cinema/admin/addTicket"));
             }
         }
+        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
+        ModelAndView mav = new ModelAndView("addTicket");
+        mav.addObject("filmshowList", filmshowList);
         return mav;
     }
 
-    @RequestMapping(value = "/admin/addTicketAll", method = RequestMethod.POST)
-    public @ResponseBody ModelAndView addTicketAll(@RequestBody Ticket ticket, List<Seat> filteredSeatList) throws Exception
+    @RequestMapping("/admin/addTicketAll")
+    public ModelAndView addTicketAll(@ModelAttribute Ticket ticket, List<Seat> filteredSeatList) throws Exception
     {
-        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
-        ModelAndView mav = new ModelAndView("addTicket");
-        mav.addObject("filmshowListJson", filmshowList);
         if(ticket != null && ticket.getFilmshow() != null
             && ticket.getSeat() != null && ticket.getPrice() != null
                 && filteredSeatList != null)
@@ -69,34 +77,44 @@ public class TicketController
                 ticket.setSeat(s);
                 ticketService.create(ticket);
             }
+            return new ModelAndView(new RedirectView("/cinema/admin/addTicketAll"));
         }
+        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
+        ModelAndView mav = new ModelAndView("addTicket");
+        mav.addObject("filmshowList", filmshowList);
         return mav;
     }
 
 //    @RequestMapping("/admin/updateTicket")
 
-    @RequestMapping("/admin/deleteTicket/{ticketId}")
-    public @ResponseBody ModelAndView deleteTicket(@PathVariable int ticketId) throws Exception
+    @RequestMapping("/admin/deleteTicket")
+    public ModelAndView deleteTicket(@ModelAttribute Ticket ticket, HttpServletResponse response) throws Exception
     {
-        List<Ticket> ticketList = ticketService.getTicketAll();
-        Ticket ticket = ticketService.getTicketById(ticketId);
-        if(ticket != null)
+        try
         {
-            ticketService.delete(ticket);
+            if(ticket != null && ticket.getFilmshow() != null
+                    && ticket.getSeat() != null && ticket.getPrice() != null)
+            {
+                ticketService.delete(ticket);
+            }
+            List<Ticket> ticketList = ticketService.getTicketAll();
+            return new ModelAndView("deleteTicket", "ticketList", ticketList);
         }
-        return new ModelAndView("deleteTicket", "ticketListJson", ticketList);
+        catch(Exception e)
+        {
+            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+            e.printStackTrace();
+            return null;
+        }
     }
 
     @RequestMapping("/admin/ticketList")
-    public @ResponseBody ModelAndView listTickets(@RequestBody Filmshow filmshow) throws Exception
+    public ModelAndView listTickets(@ModelAttribute Filmshow filmshow) throws Exception
     {
         List<Ticket> ticketList = ticketService.getTicketAll();
-        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
-        ModelAndView mav = new ModelAndView("ticketList");
-        mav.addObject("ticketListJson", ticketList);
-        mav.addObject("filmshowListJson", filmshowList);
-        List<Ticket> filteredList = new LinkedList<>();
-        if(ticketList != null && filmshow != null)
+        List<Ticket> filteredList = new ArrayList<>();
+        if(ticketList != null && filmshow != null && filmshow.getFilm() != null
+                && filmshow.getHall() != null && filmshow.getDateTime() != null)
         {
             for(Ticket t : ticketList)
             {
@@ -105,10 +123,54 @@ public class TicketController
                     filteredList.add(t);
                 }
             }
-            mav.addObject("filteredListJson", filteredList);
         }
+        List<Filmshow> filmshowList = filmshowService.getFilmshowAll();
+        ModelAndView mav = new ModelAndView("ticketList");
+        mav.addObject("ticketList", ticketList);
+        mav.addObject("filmshowList", filmshowList);
+        mav.addObject("filteredList", filteredList);
         return mav;
     }
 
+    @RequestMapping("/admin/seatsFilter/{filmshowId}")
+    public Model filterSeats(@PathVariable int filmshowId, Model model) throws Exception
+    {
+        List<Seat> seatList = seatService.getSeatAll();
+        List<Ticket> ticketList = ticketService.getTicketAll();
+        boolean seatFree;
+        Filmshow filmshow = filmshowService.getFilmshowById(filmshowId);
+        List<Seat> filteredSeatList = new ArrayList<>();
+        if(filmshow != null && filmshow.getFilm() != null
+                && filmshow.getHall() != null && filmshow.getDateTime() != null)
+        {
+            for(Seat s : seatList)
+            {
+                seatFree = true;
+                if(s.getHall().equals(filmshow.getHall()))
+                {
+                    for(Ticket t : ticketList)
+                    {
+                        if(t.getFilmshow().equals(filmshow) && t.getSeat().equals(s))
+                        {
+                            seatFree = false;
+                            break;
+                        }
+                    }
+                    if(seatFree)
+                    {
+                        filteredSeatList.add(s);
+                    }
+                }
+            }
+        }
+        model.addAttribute("filteredSeatList", filteredSeatList);
+        return model;
+    }
 
+    @InitBinder
+    public void initBinder(WebDataBinder binder)
+    {
+        binder.registerCustomEditor(Ticket.class, ticketEditor);
+        binder.registerCustomEditor(Filmshow.class, filmshowEditor);
+    }
 }
